Task_1_3 ready to optimize

+++ Week3/task1_3_runner.py
@@ -1,4 +1,5 @@
 from numpy import save
+from sympy import im
 from torch import gt
 from utils import save_frames, reset_folder
 from pathlib import Path
@@ -16,36 +17,36 @@ command = 'python TrackEval/scripts/run_mot_challenge.py --BENCHMARK custom --SP
 
 if __name__ == '__main__':
 
-    video = True
+    video = False
     sequence_path = Path('../OptionalTaskW2Dataset/train/S03/c010/') # if this is the path of some frames put video False
+    precomputed = True
 
     video_path = sequence_path / 'vdo.avi'
     path_output = './tracking_yolo_sort'
-
-    reset_folder(path_output)
-
-    video_frames_tracking = os.path.join(path_output, 'frames')
-
-    if video:
-        save_frames(str(video_path), video_frames_tracking)
-    
-    else:
-        video_frames_tracking = video_path
-
-    # get detections
     yolo_runs = Path('../yolov9/runs/detect/exp/labels/')
 
-    try:
-        shutil.rmtree('../yolov9/runs/detect/exp/')
-    except:
-        pass
+    video_frames_tracking = os.path.join(path_output, 'frames')
+    
+    if video:
+        reset_folder(path_output)
+        save_frames(str(video_path), video_frames_tracking)
+
+    # get detections
 
     # Execute yolov9
-    os.system(f'python ../yolov9/detect.py --save-txt --save-conf --classes 0 1 2 3 5 7 --weights ../yolov9/weights/YOLOv9StrategyCWeights.pt --conf 0.5 --source {video_frames_tracking} --device 0')
-    
+    if not precomputed:
+        try:
+            shutil.rmtree('../yolov9/runs/detect/exp/')
+        except:
+            pass
+
+        os.system(f'python3 ../yolov9/detect.py --save-txt --save-conf --classes 0 1 2 3 5 7 --weights ../yolov9/weights/YOLOv9StrategyCWeights.pt --conf 0.5 --source {video_frames_tracking} --device 0')
+
     det_path = str(yolo_runs)
+
+    img_path = os.path.join(video_frames_tracking, 'frame00000.png')
     
-    image = cv2.imread(os.path.join(video_frames_tracking, 'frame00000.png'))
+    image = cv2.imread(img_path, cv2.IMREAD_COLOR)
     
     h, w, _ = image.shape
     files = os.listdir(video_frames_tracking)
@@ -53,4 +54,19 @@ if __name__ == '__main__':
     optical_flow = "pyflow"
     type_combi = "median"
     
-    tracking = Tracking(det_path = str(yolo_runs), path_imgs = video_frames_tracking, path_output = path_output, w = w, h = h,num_files = num_files, display = False, ini_0 = True, classes = [0,1,2,3,5,7], conf_thr = 0.4, optical_flow = optical_flow, type_combi=type_combi)
+    tracking = Tracking(
+        pathDets = str(yolo_runs), 
+        pathImgs = video_frames_tracking, 
+        pathOutput = path_output, 
+        w = w, 
+        h = h,
+        num_files = num_files, 
+        display = False, 
+        ini_0 = True, 
+        classes = [0,1,2,3,5,7], 
+        conf_thr = 0.4, 
+        optical_flow = optical_flow, 
+        type_combi=type_combi
+    )
+
+    tracking.SORT_OF()
